requirements.txt
affine==2.3.0
attrs==19.3.0
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
Fiona==1.8.13.post1
geopandas==0.7.0
munch==2.5.0
numpy==1.18.2
pandas==1.0.3
Pillow==7.0.0
pkg-resources==0.0.0
pyparsing==2.4.6
pyproj==2.6.0
python-dateutil==2.8.1
pytz==2019.3
rasterio==1.1.3
Rtree==0.8.3
Shapely==1.7.0
six==1.14.0
snuggs==1.4.7
tqdm==4.44.1
